Add eg how to set a const static var of pure virt
In this example file, it is shown how you can set a constant static
variable of a pure virtual function, which serves as a configuration parameter for all derived
classes. The trick here is to initialize it prior to the main function.
Then, all derived classes have access to this common parameter.